+++ packages/web/src/lib/utils.ts
@@ -34,7 +34,7 @@ export const createPathWithQueryParams = (path: string, ...queryParams: [string,
 type CodeHostInfo = {
     type: "github" | "gitlab" | "gitea" | "gerrit";
     displayName: string;
-    costHostName: string;
+    codeHostName: string;
     repoLink: string;
     icon: string;
     iconClassName?: string;
@@ -57,7 +57,7 @@ export const getRepoCodeHostInfo = (repo?: Repository): CodeHostInfo | undefined
             return {
                 type: "github",
                 displayName: displayName,
-                costHostName: "GitHub",
+                codeHostName: "GitHub",
                 repoLink: repo.URL,
                 icon: githubLogo,
                 iconClassName: "dark:invert",
@@ -66,7 +66,7 @@ export const getRepoCodeHostInfo = (repo?: Repository): CodeHostInfo | undefined
             return {
                 type: "gitlab",
                 displayName: displayName,
-                costHostName: "GitLab",
+                codeHostName: "GitLab",
                 repoLink: repo.URL,
                 icon: gitlabLogo,
             }
@@ -74,7 +74,7 @@ export const getRepoCodeHostInfo = (repo?: Repository): CodeHostInfo | undefined
             return {
                 type: "gitea",
                 displayName: displayName,
-                costHostName: "Gitea",
+                codeHostName: "Gitea",
                 repoLink: repo.URL,
                 icon: giteaLogo,
             }
@@ -82,7 +82,7 @@ export const getRepoCodeHostInfo = (repo?: Repository): CodeHostInfo | undefined
             return {
                 type: "gerrit",
                 displayName: displayName,
-                costHostName: "Gerrit",
+                codeHostName: "Gerrit",
                 repoLink: repo.URL,
                 icon: gerritLogo,
             }